% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{data_load}
\alias{data_load}
\title{Load Markdown Files from the RWeekly GitHub Repository}
\usage{
data_load(
  repo_name = "rweekly/rweekly.org",
  branch_name = "gh-pages",
  skip_pattern = "2016|2017"
)
}
\arguments{
\item{repo_name}{Character. Name of the GitHub repository in the format "user/repo".
Default is "rweekly/rweekly.org".}

\item{branch_name}{Character. Name of the branch in the GitHub repository to fetch
the files from. Default is "gh-pages".}

\item{skip_pattern}{Character. A regular expression pattern specifying which files
to exclude based on their names. Default is "2016|2017", which excludes files
with names containing "2016" or "2017".}
}
\value{
A tibble with YAML metadata and content from the fetched markdown files,
where each section of content is a separate column.
}
\description{
This function fetches markdown files from a specified GitHub repository and branch,
filters out files based on given patterns, processes the files to extract YAML
metadata and content, and then reshapes the data into a wider format.
}
\examples{
\dontrun{
  data <- data_load()
  head(data)
}
}
