% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gh_mdfiles.R
\name{get_gh_mdfiles}
\alias{get_gh_mdfiles}
\title{Get list of markdown file Github URLs from Repo}
\usage{
get_gh_mdfiles(repo_name, branch_name, filter_pattern = "\\\\.md$")
}
\arguments{
\item{repo_name}{Github repository name with org or username included}

\item{branch_name}{Github repository brach name}

\item{filter_pattern}{regex pattern to filter file list by, '\\.md$' by default for markdown files}
}
\value{
character vector of Github URLs
}
\description{
get_gh_mdfiles() retrieves sourceable R code URLs from a Github repo
}
\examples{
\dontrun{
if(interactive()){
 # get_gh_mdfiles("rweekly/rweekly.org", "gh-pages")
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{stop_for_status}}
\code{\link[glue]{glue}}
\code{\link[stringr]{str_subset}}, \code{\link[stringr]{str_replace}}
}
