% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_md_to_list.R
\name{read_md_to_list}
\alias{read_md_to_list}
\title{Read and Parse Markdown File Content}
\usage{
read_md_to_list(file_path)
}
\arguments{
\item{file_path}{Character string representing the path to the markdown file.}
}
\value{
A list containing two named elements:
\itemize{
\item \code{yaml}: A list containing the parsed YAML metadata, where each name-value pair corresponds to an entry in the YAML.
\item \code{content}: A list of sections from the markdown content, where the names of the list elements are the section headers and the values are lists of the lines (content) belonging to those sections.
}
}
\description{
This function reads the content of a markdown file, extracts the YAML metadata
and sections denoted by "###" headers, and returns them in a list format.
}
\examples{
\dontrun{
  md_data <- read_md_to_list("https://raw.githubusercontent.com/rweekly/rweekly.org/gh-pages/_posts/2023-09-25-2023-W39.md")
  print(md_data$yaml)
  print(md_data$content$'Section Name')
}

}
