% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rweekly_data}
\alias{rweekly_data}
\title{RWeekly Releases data}
\format{
A data frame with over 30 columns
\describe{
\item{title}{Release title}
\item{description}{Release description}
\item{image}{Main image path}
}
}
\source{
\url{https://github.com/rweekly/rweekly.org/}
}
\usage{
rweekly_data
}
\description{
R weekly provides weekly updates from the R community. Updates are in releases published on \code{rweekly.org}.
}
\keyword{datasets}
